requirements.txt
aiogram==3.27.0
httpx==0.27.2
python-dotenv==1.2.2
